+++ src/incydr/__init__.py
@@ -3,12 +3,20 @@
 # SPDX-License-Identifier: MIT
 from . import enums
 from . import models
+from _incydr_sdk import exceptions
 from _incydr_sdk.__version__ import __version__
 from _incydr_sdk.core.client import Client
 from _incydr_sdk.queries.alerts import AlertQuery
 from _incydr_sdk.queries.file_events import EventQuery
 
-__all__ = ["__version__", "Client", "AlertQuery", "EventQuery", "models", "enums"]
+__all__ = [
+    "__version__",
+    "Client",
+    "AlertQuery",
+    "EventQuery",
+    "models",
+    "exceptions",
+]
 
 __locals = locals()
 for __name in __all__: