+++ tests/conftest.py
@@ -5,7 +5,6 @@ import pytest
 from click.testing import CliRunner
 from pytest_httpserver import HTTPServer
 
-
 TEST_SERVER_ADDRESS = "127.0.0.1:8042"
 TEST_HOST = f"http://{TEST_SERVER_ADDRESS}"